Remove CameraStage and rename CameraHandler to CameraDragListener
- now adds a CameraDragListener to stage instead of using a subclass of stage to handle camera movement
