+++ src/main/java/inf112/roborally/game/gui/Hud.java
@@ -4,7 +4,6 @@ import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.utils.viewport.FitViewport;
-import inf112.roborally.game.board.CameraStage;
 import inf112.roborally.game.objects.Player;
 
 public class Hud {
@@ -20,7 +19,7 @@ public class Hud {
         camera.update();
         viewport = new FitViewport(1920, 1080, camera);
         batch = new SpriteBatch();
-        cardsInHandDisplay = new CardsInHandDisplay(player, new CameraStage(viewport, batch));
+        cardsInHandDisplay = new CardsInHandDisplay(player, new Stage(viewport, batch));
         programRegisterDisplay = new ProgramRegisterDisplay(player, camera);
         batch.setProjectionMatrix(cardsInHandDisplay.stage.getCamera().combined);
     }
